Refactor - used layouts for common components
