+++ app.vue
@@ -1,9 +1,7 @@
 <template>
   <NuxtLayout>
-    <v-app>
-      <v-main>
-        <NuxtPage />
-      </v-main>
-    </v-app>
+    <v-main>
+      <NuxtPage />
+    </v-main>
   </NuxtLayout>
 </template>
