+++ helpers/socials.ts
@@ -1,0 +1,14 @@
+export const icons = [
+  {
+    name: 'mdi-facebook',
+    url: 'https://www.facebook.com/Politechnika.Lodzka/?locale=pl_PL',
+  },
+  {
+    name: 'mdi-twitter',
+    url: 'https://twitter.com/p_lodz_pl',
+  },
+  {
+    name: 'mdi-instagram',
+    url: 'https://www.instagram.com/Politechnika_Lodzka/',
+  },
+]
